Get, post and delete working

+++ src/db.ts
@@ -1,5 +1,8 @@
 export interface todo {
   // sketch out interface here
+  text: string;
+  createdAt: number;
+  completed: boolean;
 }
 
 export interface todoWithId extends todo {
@@ -18,14 +21,27 @@ let idCounter = 0;
  * @returns the created todos
  */
 export const addDummyTodos = (n: number): todoWithId[] => {
-  const createdSignatures: todoWithId[] = [];
+  const createdTodos: todoWithId[] = [];
   for (let count = 0; count < n; count++) {
-    const createdSignature = addTodo({
-      // possibly add some generated data here
+    const createdTodo = addTodo({
+      text: [
+        "Feed the cat",
+        "Water the plants",
+        "Put the bins out",
+        "Contemplate life",
+        "Tidy up",
+        "Walk the dog",
+        "Write a todo list",
+        "Book a holiday",
+        "Write a shopping list",
+        "Do the shopping",
+      ][Math.floor(Math.random() * 10)],
+      createdAt: +new Date(),
+      completed: false,
     });
-    createdSignatures.push(createdSignature);
+    createdTodos.push(createdTodo);
   }
-  return createdSignatures;
+  return createdTodos;
 };
 
 /**
@@ -35,9 +51,12 @@ export const addDummyTodos = (n: number): todoWithId[] => {
  * @returns the todo added (with a newly created id)
  */
 export const addTodo = (data: todo): todoWithId => {
+  const { text } = data;
   const newEntry: todoWithId = {
     id: ++idCounter,
-    ...data,
+    text,
+    createdAt: +new Date(),
+    completed: false,
   };
   db.push(newEntry);
   return newEntry;
@@ -50,14 +69,15 @@ export const addTodo = (data: todo): todoWithId => {
  * @returns the deleted database todo (if originally located),
  *  otherwise the string `"not found"`
  */
-export const deleteTodoById = (id: number): todoWithId | "not found" => {
+export const deleteTodoById = (id: number): todoWithId | "Not found" => {
   const idxToDeleteAt = findIndexOfTodoById(id);
-  if (typeof idxToDeleteAt === "number") {
+  // console.log(id, idxToDeleteAt);
+  if (typeof idxToDeleteAt === "number" && idxToDeleteAt >= 0) {
     const todoToDelete = getTodoById(id);
     db.splice(idxToDeleteAt, 1); // .splice can delete from an array
     return todoToDelete;
   } else {
-    return "not found";
+    return "Not found";
   }
 };
 
@@ -68,13 +88,13 @@ export const deleteTodoById = (id: number): todoWithId | "not found" => {
  * @returns the index of the matching database todo,
  *  otherwise the string `"not found"`
  */
-const findIndexOfTodoById = (id: number): number | "not found" => {
+const findIndexOfTodoById = (id: number): number | "Not found" => {
   const matchingIdx = db.findIndex((entry) => entry.id === id);
   // .findIndex returns -1 if not located
-  if (matchingIdx) {
+  if (matchingIdx >= 0) {
     return matchingIdx;
   } else {
-    return "not found";
+    return "Not found";
   }
 };
 
@@ -93,12 +113,12 @@ export const getAllTodos = (): todoWithId[] => {
  * @returns the located database todo (if found),
  *  otherwise the string `"not found"`
  */
-export const getTodoById = (id: number): todoWithId | "not found" => {
+export const getTodoById = (id: number): todoWithId | "Not found" => {
   const maybeEntry = db.find((entry) => entry.id === id);
   if (maybeEntry) {
     return maybeEntry;
   } else {
-    return "not found";
+    return "Not found";
   }
 };
 
@@ -114,12 +134,12 @@ export const getTodoById = (id: number): todoWithId | "not found" => {
 export const updateTodoById = (
   id: number,
   newData: Partial<todo>
-): todoWithId | "not found" => {
+): todoWithId | "Not found" => {
   const idxOfEntry = findIndexOfTodoById(id);
   // type guard against "not found"
   if (typeof idxOfEntry === "number") {
     return Object.assign(db[idxOfEntry], newData);
   } else {
-    return "not found";
+    return "Not found";
   }
 };
